Instead of throwing the message into a black hole when we are rate limited, wait "retry_after" milliseconds until trying to send again.
This also adds a "blocking" state, where the second parameter of the
TemmieWebhook constructor sets a "blocking" (or not) state... By default
it doesn't block.

When blockMainThread = true: A Runnable is created and r.run();
When blockMainThread = false: A Runnable is created and a Thread is
created with the Runnable, simple, huh?

I hope this isn't bad code because... I don't know, feels like it isn't
good, but hey, it WORKS(TM)